Refactor angle calculation module to unify MOA and MRAD handling under a single struct and remove deprecated files
